ANDROID: GKI: allow abi break in struct io_ring_ctx
In commit 46b1b3d81a7e ("io_uring/kbuf: use vm_insert_pages() for
mmap'ed pbuf ring"), a structure was removed from struct io_ring_ctx.
This structure is private to the io_uring core code, but exposed as a
pointer to it is in a global structure used elsewhere.  Because of this,
the abi checker, and CRC checker, notices the change and causes an abi
break to be reported.

Fix thsi up by allowing this break to be updated to the .stg file and
document it as being ok.

Change-Id: I8c4c182112f521910399be07285c21d3aac70711
Fixes: 46b1b3d81a7e ("io_uring/kbuf: use vm_insert_pages() for mmap'ed pbuf ring")
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>